Abstract a common WatchTileService.kt This provides the watch tied to QS tiles, as well as simplified coroutine support
